+++ eshop-service/src/main/java/com/eshop/eshopservice/service/InventoryService.java
@@ -85,7 +85,7 @@ public class InventoryService implements InventoryServiceInterface {
 	 * @throws InventoryProductException
 	 */
 	@Override
-	public InventoryProductDTO updateInventoryProducts(int inventoryProductID, InventoryProductDTO inventoryProductDTOObject) 
+	public InventoryProductDTO updateInventoryProducts(long inventoryProductID, InventoryProductDTO inventoryProductDTOObject) 
 			throws InventoryProductException {
 		InventoryProduct inventoryProductRetrieveObject = inventoryProductRepository.findById(inventoryProductID).
 				orElseThrow(() -> new InventoryProductException(messageSource.getMessage("InventoryProductNotFound", null, LocaleContextHolder.getLocale())));
@@ -114,7 +114,7 @@ public class InventoryService implements InventoryServiceInterface {
 	 * @throws InventoryProductException
 	 */
 	@Override
-	public void removeInventoryProduct(int inventoryProductID) throws InventoryProductException {
+	public void removeInventoryProduct(long inventoryProductID) throws InventoryProductException {
 		if(!inventoryProductRepository.existsById(inventoryProductID)) {
 			throw new InventoryProductException(messageSource.getMessage("InventoryProductNotFound", null, LocaleContextHolder.getLocale())); 
 		}
